Move PDESampler object to library (rather than sampler.cpp example)
