+++ examples/sampler.cpp
@@ -106,151 +106,6 @@ void Visualize(const mfem::Vector& sol,
     MPI_Barrier(pmesh->GetComm());
 };
 
-/// scalar normal distribution
-class NormalDistribution
-{
-public:
-    NormalDistribution(double mean = 0.0, double stddev = 1.0, int seed = 0);
-    double Sample();
-private:
-    std::mt19937 generator_;
-    std::normal_distribution<double> dist_;
-};
-
-NormalDistribution::NormalDistribution(double mean, double stddev, int seed)
-    :
-    generator_(seed),
-    dist_(mean, stddev)
-{
-}
-
-double NormalDistribution::Sample()
-{
-    double out = dist_(generator_);
-    return out;
-}
-
-/**
-   For now we return the log of the permeability field,
-   at some point this class (or its caller?) should
-   exponentiate.
-*/
-class PDESampler
-{
-public:
-    /**
-       @todo cell_volume should be potentially spatially-varying
-    */
-    PDESampler(const FiniteVolumeUpscale& fvupscale,
-               int fine_vector_size, int dimension, double cell_volume,
-               double kappa, int seed);
-    ~PDESampler();
-
-    /// Draw white noise on fine level
-    void Sample();
-
-    /// Draw white noise on coarse level
-    void CoarseSample();
-
-    /// Solve PDE with white-noise RHS to find fine coefficient
-    mfem::Vector& GetFineCoefficient();
-
-    /// Solve PDE with white-noise RHS to find coarse coeffiicent
-    mfem::Vector& GetCoarseCoefficient();
-
-private:
-    enum State
-    {
-        NO_SAMPLE,
-        FINE_SAMPLE,
-        COARSE_SAMPLE
-    };
-
-    const FiniteVolumeUpscale& fvupscale_;
-    NormalDistribution normal_distribution_;
-    int fine_vector_size_;
-    double cell_volume_;
-    double scalar_g_;
-    State current_state_;
-
-    /// all these vectors live in the pressure / vertex space
-    mfem::Vector rhs_fine_;
-    mfem::Vector rhs_coarse_;
-    mfem::Vector coefficient_fine_;
-    mfem::Vector coefficient_coarse_;
-};
-
-PDESampler::PDESampler(const FiniteVolumeUpscale& fvupscale,
-                       int fine_vector_size, int dimension, double cell_volume,
-                       double kappa, int seed)
-    :
-    fvupscale_(fvupscale),
-    normal_distribution_(0.0, 1.0, seed),
-    fine_vector_size_(fine_vector_size),
-    cell_volume_(cell_volume),
-    current_state_(NO_SAMPLE)
-{
-    rhs_fine_.SetSize(fine_vector_size_);
-    coefficient_fine_.SetSize(fine_vector_size_);
-    rhs_coarse_ = fvupscale_.GetCoarseVector();
-    coefficient_coarse_ = fvupscale_.GetCoarseVector();
-
-    double nu_parameter;
-    if (dimension == 2)
-        nu_parameter = 1.0;
-    else
-        nu_parameter = 0.5;
-    double ddim = static_cast<double>(dimension);
-    scalar_g_ = std::pow(4.0 * M_PI, ddim / 4.0) * std::pow(kappa, nu_parameter) *
-                std::sqrt( tgamma(nu_parameter + ddim / 2.0) / tgamma(nu_parameter) );
-}
-
-PDESampler::~PDESampler()
-{
-}
-
-/// @todo cell_volume should be variable rather than constant
-void PDESampler::Sample()
-{
-    current_state_ = FINE_SAMPLE;
-
-    // construct white noise right-hand side
-    // (cell_volume is supposed to represent fine-grid W_h)
-    for (int i = 0; i < fine_vector_size_; ++i)
-    {
-        rhs_fine_(i) = scalar_g_ * std::sqrt(cell_volume_) *
-                       normal_distribution_.Sample();
-    }
-}
-
-void PDESampler::CoarseSample()
-{
-    current_state_ = COARSE_SAMPLE;
-    MFEM_ASSERT(false, "Not implemented!");
-}
-
-mfem::Vector& PDESampler::GetFineCoefficient()
-{
-    MFEM_ASSERT(current_state_ == FINE_SAMPLE,
-                "PDESampler object in wrong state (call Sample() first)!");
-
-    fvupscale_.SolveFine(rhs_fine_, coefficient_fine_);
-    return coefficient_fine_;
-}
-
-mfem::Vector& PDESampler::GetCoarseCoefficient()
-{
-    MFEM_ASSERT(current_state_ == FINE_SAMPLE ||
-                current_state_ == COARSE_SAMPLE,
-                "PDESampler object in wrong state (call Sample() first)!");
-
-    if (current_state_ == FINE_SAMPLE)
-        fvupscale_.Restrict(rhs_fine_, rhs_coarse_);
-    fvupscale_.SolveCoarse(rhs_coarse_, coefficient_coarse_);
-    coefficient_coarse_ *= -1.0; // ??
-    return coefficient_coarse_;
-}
-
 int main(int argc, char* argv[])
 {
     int num_procs, myid;
