+++ src/CMakeLists.txt
@@ -23,5 +23,5 @@ add_library(smoothg LocalMixedGraphSpectralTargets.cpp GraphCoarsenBuilder.cpp
   sharedentitycommunication.cpp GraphTopology.cpp Mixed_GL_Coarsener.cpp
   SpectralAMG_MGL_Coarsener.cpp MetisGraphPartitioner.cpp MatrixUtilities.cpp
   MixedMatrix.cpp LocalEigenSolver.cpp GraphGenerator.cpp GraphUpscale.cpp FiniteVolumeUpscale.cpp
-  FiniteVolumeMLMC.cpp Upscale.cpp MixedLaplacianSolver.cpp)
+  FiniteVolumeMLMC.cpp Upscale.cpp MixedLaplacianSolver.cpp Sampler.cpp)
